Separate handler's package with client
