+++ src/thread/handler.go
@@ -1,4 +1,4 @@
-package client
+package thread
 
 import (
 	"encoding/json"
@@ -10,15 +10,15 @@ import (
 	"time"
 )
 
-type handler struct {
+type Handler struct {
 	conn     net.Conn
 	value    []string
-	schedule sync.WaitGroup
-	send     chan string
+	Schedule sync.WaitGroup
+	Send     chan string
 	err      error
 }
 
-func (e *handler) Create(serverAddr string, value []string) {
+func (e *Handler) Create(serverAddr string, value []string) {
 	e.conn, e.err = net.Dial("tcp", serverAddr)
 	e.value = value
 	if e.err != nil {
@@ -26,26 +26,26 @@ func (e *handler) Create(serverAddr string, value []string) {
 	}
 }
 
-func (e *handler) requestMaker() {
+func (e *Handler) RequestMaker() {
 	log.Printf("Logger: handler.requestMaker() handler=%v", e)
 	defer log.Printf("Logger: handler.requestMaker() handler=%v", e)
 
 	for {
 		select {
-		case msg := <-e.send:
+		case msg := <-e.Send:
 			ch := make(chan string, 2)
 			time.Sleep(100 * time.Millisecond)
 			if err := e.MakeRequest(msg, ch); nil != err {
 				log.Printf("failed request err: %v", err)
 			}
 			e.ListenResponse(ch)
-			e.schedule.Done()
+			e.Schedule.Done()
 			break
 		}
 	}
 }
 
-func (e *handler) MakeRequest(Message string, ch chan string) error {
+func (e *Handler) MakeRequest(Message string, ch chan string) error {
 
 	lock := &sync.Mutex{}
 	lock.Lock()
@@ -64,7 +64,7 @@ func (e *handler) MakeRequest(Message string, ch chan string) error {
 	return nil
 }
 
-func (e *handler) ListenResponse(ch chan string) {
+func (e *Handler) ListenResponse(ch chan string) {
 
 	buf := make([]byte, 0, 16384)
 	tmp := make([]byte, 256)
